Extract `AbstractPrecompiledScriptPluginTest` test superclass
